Rename the effect component
