+++ components/learn/TheLearnHeader.vue
@@ -2,7 +2,7 @@
   <ThePageHeader class="the-learn-header">
     Start your path towards
     <br>learning
-    <TypedWord
+    <TypewriterEffect
       :values="['Qiskit', 'Python', 'Quantum']"
     />
   </ThePageHeader>
@@ -12,8 +12,8 @@
 import Vue from 'vue'
 import { Component } from 'vue-property-decorator'
 import ThePageHeader from '~/components/ui/ThePageHeader.vue'
-import TypedWord from '~/components/ui/TypedWord.vue'
+import TypewriterEffect from '~/components/ui/TypewriterEffect.vue'
 
-@Component({ components: { ThePageHeader, TypedWord } })
+@Component({ components: { ThePageHeader, TypewriterEffect } })
 export default class extends Vue { }
 </script>
